apply path aliases refactor
