+++ src/pages/landelijk/verdenkingen-huisartsen.tsx
@@ -1,17 +1,15 @@
+import siteText from '~/locale/index';
+import { VerdenkingenHuisartsen } from '~/types/data.d';
+import getNlData, { INationalData } from '~/static-props/nl-data';
+
 import { ContentHeader } from '~/components/layout/Content';
 import { FCWithLayout } from '~/components/layout';
 import { getNationalLayout } from '~/components/layout/NationalLayout';
-import { NationalPractitionerLineChart } from '~/components/charts/index';
-
+import { NationalPractitionerLineChart } from '~/components/lineChart/nationalPractitionerLineChart';
 import { SuspectedPatientsBarScale } from '~/components/landelijk/suspected-patients-barscale';
 
 import Arts from '~/assets/arts.svg';
 import { formatNumber } from '~/utils/formatNumber';
-
-import siteText from '~/locale/index';
-
-import { VerdenkingenHuisartsen } from '~/types/data.d';
-import getNlData, { INationalData } from '~/static-props/nl-data';
 
 const text: typeof siteText.verdenkingen_huisartsen =
   siteText.verdenkingen_huisartsen;
